+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Notable changes to this project.
 - add `daily_submissions_performances`
 - remove v1 staking
 - remove `get_staking_cutoff` - no longer relevant
+- `get_leaderboard` renamed to `get_v1_leaderboard`
 
 ## [1.6.2] - 2019-07-31
 - remove phone number and bonus fetching (#16)